feat: add render resize function
